+++ cli/src/generator/prelude.rs
@@ -1,0 +1,5 @@
+pub use prisma_client_rust_sdk::prisma_datamodel::dml;
+pub use prisma_client_rust_sdk::*;
+pub use proc_macro2::*;
+pub use quote::*;
+pub use syn::Ident;
